+++ packages/library-sync/src/commonMain/kotlin/io/realm/kotlin/mongodb/AppConfiguration.kt
@@ -35,6 +35,7 @@ import io.realm.kotlin.log.LogLevel
 import io.realm.kotlin.log.RealmLogger
 import io.realm.kotlin.mongodb.internal.AppConfigurationImpl
 import io.realm.kotlin.mongodb.internal.KtorNetworkTransport
+import io.realm.kotlin.mongodb.internal.LogObfuscatorImpl
 import io.realm.kotlin.mongodb.sync.SyncConfiguration
 import kotlinx.coroutines.CoroutineDispatcher
 import org.mongodb.kbson.ExperimentalKSerializerApi
@@ -75,6 +76,12 @@ public interface AppConfiguration {
      */
     @OptIn(ExperimentalKSerializerApi::class)
     public val ejson: EJson
+
+    /**
+     * The configured [HttpLogObfuscator] for this app. If this property returns `null` no
+     * obfuscator is being used.
+     */
+    public val httpLogObfuscator: HttpLogObfuscator?
 
     public companion object {
         /**
@@ -120,6 +127,7 @@ public interface AppConfiguration {
         private var appVersion: String? = null
         @OptIn(ExperimentalKSerializerApi::class)
         private var ejson: EJson = EJson
+        private var httpLogObfuscator: HttpLogObfuscator? = LogObfuscatorImpl
 
         /**
          * Sets the encryption key used to encrypt the user metadata Realm only. Individual
@@ -240,6 +248,20 @@ public interface AppConfiguration {
         }
 
         /**
+         * Sets the a [HttpLogObfuscator] used to keep sensitive information in HTTP requests from
+         * being displayed in the log. Logs containing tokens, passwords or custom function
+         * arguments and the result of computing these will be obfuscated by default. Logs will not
+         * be obfuscated if the value is set to `null`.
+         *
+         * @param httpLogObfuscator the HTTP log obfuscator to be used or `null` if obfuscation
+         * should be disabled.
+         * @return the Builder instance used.
+         */
+        public fun httpLogObfuscator(httpLogObfuscator: HttpLogObfuscator?): Builder = apply {
+            this.httpLogObfuscator = httpLogObfuscator
+        }
+
+        /**
          * TODO
          */
         @ExperimentalKSerializerApi
@@ -291,17 +313,23 @@ public interface AppConfiguration {
             }
 
             val networkTransport: () -> NetworkTransport = {
+                val logger: Logger? = if (logLevel <= LogLevel.DEBUG) {
+                    object : Logger {
+                        override fun log(message: String) {
+                            val obfuscatedMessage = httpLogObfuscator?.obfuscate(message)
+                            appLogger.debug(obfuscatedMessage ?: message)
+                        }
+                    }
+                } else {
+                    null
+                }
                 networkTransport ?: KtorNetworkTransport(
                     // FIXME Add AppConfiguration.Builder option to set timeout as a Duration with default \
                     //  constant in AppConfiguration.Companion
                     //  https://github.com/realm/realm-kotlin/issues/408
-                    timeoutMs = 15000,
+                    timeoutMs = 60000,
                     dispatcherFactory = appNetworkDispatcherFactory,
-                    logger = object : Logger {
-                        override fun log(message: String) {
-                            appLogger.debug(message)
-                        }
-                    }
+                    logger = logger
                 )
             }
 
@@ -317,7 +345,8 @@ public interface AppConfiguration {
                 log = appLogger,
                 appName = appName,
                 appVersion = appVersion,
-                ejson = ejson
+                ejson = ejson,
+                httpLogObfuscator = httpLogObfuscator
             )
         }
     }